Rename rowReporter -> runReport and move to reporters/helpers/
"rowReporter" read like a reporter (a sibling of pretty/json/ndjson) when it's
actually the shared engine they delegate to. Renamed the function to `runReport`
and the file to reporters/helpers/runReport.ts — it reads as
`runReport(options, paths, { onRow, onDone })` at the call sites and no longer
masquerades as a reporter. (ReportRow + toReportRow + tests move with it.)
